# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Eco Turismo Experiencial API
  description: |
    API REST para la plataforma de reservas de alojamientos rurales y actividades de naturaleza en Colombia.
    
    ## Características
    - Autenticación JWT con RS256
    - Reservas con bloqueo optimista
    - Pagos con Stripe
    - Suscripciones con planes (basic, premium, pro)
    - Búsqueda geoespacial con radio de búsqueda
    
    ## Autenticación
    Todos los endpoints protegidos requieren un token JWT en el header:
    ```
    Authorization: Bearer <token>
    ```
  version: 1.0.0
  contact:
    name: Soporte Eco Turismo
    email: soporte@ecoturismo.com
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor de desarrollo
  - url: https://api.ecoturismo.com/api/v1
    description: Servidor de producción
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT con firma RS256
  schemas:
...
    ListingCreateRequest:
      type: object
      required:
        - title
        - type
        - price_per_unit
        - capacity
      properties:
        title:
          type: string
        type:
          type: string
          enum: [accommodation, activity]
        description:
          type: string
        price_per_unit:
          type: integer
        capacity:
          type: integer
        categories:
          type: array
          items:
            type: string
        address:
          type: string
        department:
          type: string
        municipality:
          type: string
        latitude:
          type: number
        longitude:
          type: number
...
paths:
...
  /listings:
    get:
      tags:
        - Listings
      summary: Buscar listings con filtros
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [accommodation, activity]
        - name: department
          in: query
          schema:
            type: string
        - name: municipality
          in: query
          schema:
            type: string
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: integer
            enum: [5, 10, 25, 50]
        - name: price_min
          in: query
          schema:
            type: integer
        - name: price_max
          in: query
          schema:
            type: integer
        - name: capacity
          in: query
          schema:
            type: integer
        - name: min_rating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: sort
          in: query
          schema:
            type: string
            enum: [relevance, price_asc, price_desc, rating]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        200:
          description: Lista de listings

    post:
      tags:
        - Listings
      summary: Crear nuevo listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        201:
          description: Listing creado
  /listings/{id}:
    get:
      tags:
        - Listings
      summary: Obtener detalle de un listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Detalle del listing
        404:
          description: Listing no encontrado

    put:
      tags:
        - Listings
      summary: Actualizar listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Listing actualizado

    delete:
      tags:
        - Listings
      summary: Eliminar listing (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Listing eliminado
  /listings/map:
    get:
      tags:
        - Listings
      summary: Obtener listings para mapa
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: department
          in: query
          schema:
            type: string
      responses:
        200:
          description: Lista de listings con coordenadas
  /listings/my:
    get:
      tags:
        - Listings
      summary: Obtener listings del host autenticado
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        200:
          description: Lista de listings del host
  /listings/{id}/photos:
    post:
      tags:
        - Listings
      summary: Subir fotos al listing
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photos:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        200:
          description: Fotos añadidas
...
  /reviews/listing/{listingId}:
    get:
      tags:
        - Reviews
      summary: Obtener reseñas de un listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Lista de reseñas
...
  /admin/listings:
    get:
      tags:
        - Admin
      summary: Listar publicaciones
      security:
        - bearerAuth: []
      responses:
        200:
          description: Lista de listings
...